Add files via upload
Both the IPython Notebooks contain cells of code that achieve the same thing. The varying thresholds are accounted for. However, running these two IPython Notebooks generates different extensions in their outputs. As is evident by the text in the parentheses, one generates models in ".pb" format, while the other generates ".h5". I have run both notebooks and uploaded their respective outputs here so that evaluation can be a simple plug-and-play.